Check bindgen output into the repo (#22)
The DeepSpeech C API doesn't contain any platform-specific definitions, so it's not necessary for every user of the library to re-run bindgen on their machine when building deepspeech-rs. Instead, we can run bindgen once, whenever the deepspeech submodule is bumped, check the generated rust code into this repository, and include it in the published crate.

This saves users of the crate a lot of headaches, because bindgen has a bunch of dependencies, requires certain versions of clang, and generates annoying warning messages whenever I build deepspeech locally. It also avoids some unnecessary complexity in the build script.