Generate the bindings via a dedicated crate
There are bazillion different bindgen installations,
varying in compiler version, version of dependencies,
versions of bindgen itself. If we use the binary
bindgen, it would be an endless back and forth in
minuscle differences between the installations.

With this patch, versions of the entire dependency
graph are checked into the repo. This avoids
confusion.